Refactoring mongo package to be able to include other database packages and select the tqdm.auto as base of tqdm.
