+++ dbtqdm/__init__.py
@@ -1,3 +1,2 @@
 from .db import DatabaseTqdm
-from .mongo import MongoTqdm as tqdm
-from .dbutils import connect_db
+
