+++ dbtqdm/mongo/__init__.py
@@ -1,0 +1,2 @@
+from .mongo import MongoTqdm as tqdm
+from .utils import connect_db
